Kind of a mess..

+++ SWPCCBilling.SiteHost/Models/Fee.cs
@@ -2,16 +2,16 @@
 {
     public class Fee
     {
-        public int Id { get; set; } 
+        public long Id { get; set; } 
         public string Name { get; set; }
         public string Type { get; set; }
-        public decimal? Amount { get; set; }
+        public double? Amount { get; set; }
 
         public Fee()
         {
         }
 
-        public Fee(int id, string name, string type, decimal amount)
+        public Fee(int id, string name, string type, double amount)
         {
             Id = id;
             Name = name;
